More disable user stuff...Re-vamped add user screen to be more user-friendly.  User can now add a new contact directly from that page.

+++ src/java/org/aspcfs/modules/contacts/actions/ExternalContacts.java
@@ -1181,12 +1181,19 @@ public final class ExternalContacts extends CFSModule {
     try {
       db = this.getConnection(context);
       thisContact = new Contact(db, id);
+      thisContact.checkUserAccount(db);
       htmlDialog.setRelationships(thisContact.processDependencies(db));
-              
+      
+      if (!thisContact.hasAccount()) {
         htmlDialog.setTitle("CFS: Confirm Delete");
         htmlDialog.setHeader("The object you are requesting to delete has the following dependencies within CFS:");
         htmlDialog.addButton("Delete All", "javascript:window.location.href='/ExternalContacts.do?command=DeleteContact&id=" + id + "'");
         htmlDialog.addButton("Cancel", "javascript:parent.window.close()");
+      } else {
+        htmlDialog.setHeader("This Contact cannot be deleted because it is associated with a User account.");
+        htmlDialog.addButton("OK", "javascript:parent.window.close()");
+      }        
+        
     } catch (Exception e) {
       errorMessage = e;
     } finally {
